resturered tests, switched to jest and working through failing tests
