+++ .gitignore
@@ -13,5 +13,6 @@ node_modules/
 dist/
 bin/
 tmp/
-test/test-files
+tests/test-files
 bundle/
+coverage/
