+++ jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    transform: { "^.+\\.ts?$": "ts-jest" },
+    testEnvironment: "node",
+    testRegex: "/tests/.*\\.(test|spec)?\\.(ts|tsx)$",
+    moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
+};
