--- prettierrc.yaml
@@ -1,4 +1,0 @@
-trailingComma: all
-tabWidth: 4
-singleQuote: true
-arrowParens: always
