+++ src/templates.ts
@@ -1,3 +1,5 @@
+import { MDMError } from "./commands";
+
 const templates: { [key: string]: string } = {};
 
 /**
@@ -6,6 +8,8 @@ const templates: { [key: string]: string } = {};
  * @param content The replacement string
  */
 export function new_template(name: string, content: string) {
+    if (name in templates)
+        throw new MDMError(`Template ${name} already exists`);
     templates[name] = content;
 }
