--- test/hooks.js
@@ -1,9 +1,0 @@
-const fs = require("fs");
-
-exports.mochaHooks = {
-    beforeEach(done) {
-        fs.rmSync("test/test-files", { recursive: true });
-        fs.mkdirSync("test/test-files");
-        done();
-    },
-};
