+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-    "include": ["./src/*"],
+    "include": ["./src/**/*"],
 
     "compilerOptions": {
         /* Visit https://aka.ms/tsconfig.json to read more about this file */
@@ -10,14 +10,14 @@
         "module": "commonjs" /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', 'es2020', or 'ESNext'. */,
         // "lib": [],                                   /* Specify library files to be included in the compilation. */
         "allowJs": true /* Allow javascript files to be compiled. */,
-        // "checkJs": true,                             /* Report errors in .js files. */
+        "checkJs": true /* Report errors in .js files. */,
         // "jsx": "preserve",                           /* Specify JSX code generation: 'preserve', 'react-native', 'react', 'react-jsx' or 'react-jsxdev'. */
         "declaration": true /* Generates corresponding '.d.ts' file. */,
         // "declarationMap": true,                      /* Generates a sourcemap for each corresponding '.d.ts' file. */
         "sourceMap": true /* Generates corresponding '.map' file. */,
         // "outFile": "./",                             /* Concatenate and emit output to single file. */
         "outDir": "./bundle" /* Redirect output structure to the directory. */,
-        // "rootDir": "./",                             /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
+        "rootDir": "./src" /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
         // "composite": true,                           /* Enable project compilation */
         // "tsBuildInfoFile": "./",                     /* Specify file to store incremental compilation information */
         // "removeComments": true,                      /* Do not emit comments to output. */
@@ -27,7 +27,7 @@
         // "isolatedModules": true,                     /* Transpile each file as a separate module (similar to 'ts.transpileModule'). */
 
         /* Strict Type-Checking Options */
-        // "strict": true,                                 /* Enable all strict type-checking options. */
+        // "strict": true /* Enable all strict type-checking options. */,
         // "noImplicitAny": true,                       /* Raise error on expressions and declarations with an implied 'any' type. */
         // "strictNullChecks": true,                    /* Enable strict null checks. */
         // "strictFunctionTypes": true,                 /* Enable strict checking of function types. */
